+++ CinemaKit/Sources/StartupManager.swift
@@ -73,7 +73,7 @@ public class CinemaKitStartupManager: StartupManager {
 
   private let application: UIApplication
   private let container = CKContainer.default()
-  private let userDefaults: UserDefaultsProtocol = UserDefaults.standard
+  private let userDefaults = StandardUserDefaults()
   private let migratedLibraryName: String
   private var progressHandler: ((StartupProgress) -> Void)!
 
@@ -91,11 +91,7 @@ public class CinemaKitStartupManager: StartupManager {
              type: .info,
              previousVersion.description)
       if previousVersion < currentVersion {
-        os_log("migrating from version %{public}@",
-               log: CinemaKitStartupManager.logger,
-               type: .info,
-               previousVersion.description)
-        clearPosterCache()
+        migrate(from: previousVersion)
         markCurrentVersion()
       } else if previousVersion > currentVersion {
         fatalError("going back from \(previousVersion) to \(currentVersion) is not supported -> clear app data")
@@ -113,18 +109,6 @@ public class CinemaKitStartupManager: StartupManager {
     }
     setUpDirectories()
     setUpDeviceSyncZone()
-  }
-
-  private func clearPosterCache() {
-    do {
-      os_log("clearing poster cache", log: CinemaKitStartupManager.logger, type: .default)
-      try FileManager.default.removeItem(at: CinemaKitStartupManager.posterCacheDir)
-    } catch {
-      os_log("unable to clear poster cache: %{public}@",
-             log: CinemaKitStartupManager.logger,
-             type: .fault,
-             String(describing: error))
-    }
   }
 
   private func markCurrentVersion() {
@@ -411,4 +395,41 @@ private class DefaultMovieLibraryFactory: MovieLibraryFactory {
 
 private func fail() -> Never {
   fatalError("error during startup")
+}
+
+// MARK: - migration
+
+extension CinemaKitStartupManager {
+  private func migrate(from previousVersion: AppVersion) {
+    os_log("migrating from version %{public}@",
+           log: CinemaKitStartupManager.logger,
+           type: .info,
+           previousVersion.description)
+    if previousVersion < "2.0" {
+      clearPosterCache()
+    }
+    if previousVersion < "2.0.2" {
+      renamePrimaryLibraryKey()
+    }
+  }
+
+  private func clearPosterCache() {
+    do {
+      os_log("clearing poster cache", log: CinemaKitStartupManager.logger, type: .default)
+      try FileManager.default.removeItem(at: CinemaKitStartupManager.posterCacheDir)
+    } catch {
+      os_log("unable to clear poster cache: %{public}@",
+             log: CinemaKitStartupManager.logger,
+             type: .fault,
+             String(describing: error))
+    }
+  }
+
+  private func renamePrimaryLibraryKey() {
+    if let primaryLibrary = UserDefaults.standard.string(forKey: "primaryLibrary") {
+      os_log("renaming user defaults key", log: CinemaKitStartupManager.logger, type: .default)
+      UserDefaults.standard.removeObject(forKey: "primaryLibrary")
+      UserDefaults.standard.set(primaryLibrary, forKey: "PrimaryLibrary")
+    }
+  }
 }