+++ iOS/Application/Sources/Genre/GenreListController.swift
@@ -138,6 +138,35 @@ extension GenreListController {
 
 // MARK: - Table View
 
+extension UITableView {
+  fileprivate func reloadRow(for genre: GenreListController.Genre,
+                             at indexPathProvider: @escaping () -> IndexPath?,
+                             using genreImageProvider: GenreImageProvider) {
+    guard let indexPath = indexPathProvider() else { return }
+    if let cell = cellForRow(at: indexPath) as? GenreCell {
+      configure(cell,
+                for: genre,
+                at: indexPathProvider,
+                using: genreImageProvider)
+    }
+  }
+
+  fileprivate func configure(_ cell: GenreCell,
+                             for genre: GenreListController.Genre,
+                             at indexPathProvider: @escaping () -> IndexPath?,
+                             using genreImageProvider: GenreImageProvider) {
+    cell.configure(for: genre, genreImageProvider: genreImageProvider) {
+      guard let indexPath = indexPathProvider() else { return }
+      if let cell = self.cellForRow(at: indexPath) as? GenreCell {
+        self.configure(cell,
+                       for: genre,
+                       at: indexPathProvider,
+                       using: genreImageProvider)
+      }
+    }
+  }
+}
+
 extension GenreListController: UITableViewDataSourcePrefetching {
   override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
     guard let viewModel = self.viewModel else { return 0 }
@@ -147,11 +176,13 @@ extension GenreListController: UITableViewDataSourcePrefetching {
   override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
     let cell: GenreCell = tableView.dequeueReusableCell(for: indexPath)
     let genre = viewModel[indexPath.row]
-    cell.configure(for: genre, genreImageProvider: genreImageProvider) { [weak self] in
-      guard let `self` = self else { return }
-      guard let rowIndex = self.viewModel.index(where: { $0.id == genre.id }) else { return }
-      tableView.reloadRowWithoutAnimation(at: IndexPath(row: rowIndex, section: 0))
-    }
+    tableView.configure(cell,
+                        for: genre,
+                        at: { [weak self] in
+                          self?.viewModel.firstIndex { $0.id == genre.id }
+                                         .map { IndexPath(row: $0, section: 0) }
+                        },
+                        using: genreImageProvider)
     return cell
   }
 
@@ -166,14 +197,17 @@ extension GenreListController: UITableViewDataSourcePrefetching {
   func tableView(_ tableView: UITableView, prefetchRowsAt indexPaths: [IndexPath]) {
     for indexPath in indexPaths {
       let genre = viewModel[indexPath.row]
-      if case .unknown = genre.image {
-        genre.image = .loading
-        DispatchQueue.global(qos: .background).async {
-          fetchBackdrop(for: genre, using: self.genreImageProvider) { [weak self] in
-            guard let `self` = self else { return }
-            guard let rowIndex = self.viewModel.index(where: { $0.id == genre.id }) else { return }
-            tableView.reloadRowWithoutAnimation(at: IndexPath(row: rowIndex, section: 0))
-          }
+      guard case .unknown = genre.image else { return }
+      genre.image = .loading
+      DispatchQueue.global(qos: .background).async {
+        fetchBackdrop(for: genre, using: self.genreImageProvider) { [weak self] in
+          guard let `self` = self else { return }
+          tableView.reloadRow(for: genre,
+                              at: { [weak self] in
+                                self?.viewModel.firstIndex { $0.id == genre.id }
+                                               .map { IndexPath(row: $0, section: 0) }
+                              },
+                              using: self.genreImageProvider)
         }
       }
     }
@@ -190,6 +224,7 @@ class GenreCell: UITableViewCell {
 
   override func awakeFromNib() {
     super.awakeFromNib()
+    contentView.backgroundColor = .missingArtworkBackground
     scrim = ScrimView()
     contentView.insertSubview(scrim, belowSubview: genreNameLabel)
     genreNameLabel.layer.shadowColor = UIColor.black.cgColor
@@ -218,28 +253,37 @@ class GenreCell: UITableViewCell {
       case let .available(genreImage):
         genreNameLabel.textColor = .white
         genreNameLabel.layer.shadowOpacity = 1.0
-        backdropImageView.image = genreImage
-        backdropImageView.contentMode = .scaleAspectFill
         scrim.isHidden = false
+        backdropImageView.contentMode = .scaleAspectFill
+        configureBackdrop(genreImage)
         self.activityIndicator.stopAnimating()
       case .unavailable:
         genreNameLabel.textColor = #colorLiteral(red: 0.5, green: 0.5, blue: 0.5, alpha: 1.0)
         genreNameLabel.layer.shadowOpacity = 0.0
-        backdropImageView.image = #imageLiteral(resourceName: "MissingGenreImage")
-        backdropImageView.contentMode = .center
-        backdropImageView.backgroundColor = .missingArtworkBackground
         scrim.isHidden = true
-        self.activityIndicator.stopAnimating()
+        backdropImageView.contentMode = .center
+        configureBackdrop(#imageLiteral(resourceName: "MissingGenreImage"))
+        activityIndicator.stopAnimating()
     }
   }
 
   private func configureBackdropForUnknownOrLoadingImageState() {
     genreNameLabel.textColor = #colorLiteral(red: 0.5, green: 0.5, blue: 0.5, alpha: 1.0)
     genreNameLabel.layer.shadowOpacity = 0.0
-    backdropImageView.image = nil
-    backdropImageView.backgroundColor = .missingArtworkBackground
     scrim.isHidden = true
+    configureBackdrop(nil)
     activityIndicator.startAnimating()
+  }
+
+  private func configureBackdrop(_ image: UIImage?) {
+    backdropImageView.image = image
+    if image == nil {
+      backdropImageView.alpha = 0.0
+    } else if backdropImageView.alpha < 1.0 {
+      UIView.animate(withDuration: 0.2) {
+        self.backdropImageView.alpha = 1.0
+      }
+    }
   }
 
   override func prepareForReuse() {