+++ iOS/Application/Sources/LibraryContent/SortDescriptor.swift
@@ -1,7 +1,7 @@
 import CinemaKit
 import Foundation
 
-enum SortDescriptor {
+enum SortDescriptor: String, CaseIterable {
   case title, runtime, year
 
   func makeTableViewStrategy() -> SectionSortingStrategy {
@@ -137,11 +137,13 @@ private struct RuntimeSortingStrategy: SectionSortingStrategy {
 
 private struct YearSortingStrategy: SectionSortingStrategy {
   private let unknownSymbol = "?"
+  private let currentYear = Calendar.current.component(.year, from: Date())
+  private let numberOfStandaloneYears = 10
 
   func sectionIndexTitle(for movie: Movie) -> String {
     guard let releaseDate = movie.releaseDate else { return unknownSymbol }
     let year = Calendar.current.component(.year, from: releaseDate)
-    if year < 2010 {
+    if year < currentYear - numberOfStandaloneYears {
       return String(year / 10 * 10)
     } else {
       return String(year)
@@ -162,7 +164,13 @@ private struct YearSortingStrategy: SectionSortingStrategy {
       case unknownSymbol:
         return NSLocalizedString("sort.by.year.unknownHeader", comment: "")
       default:
-        return sectionIndexTitle
+        let year = Int(sectionIndexTitle)!
+        let thresholdYear = currentYear - numberOfStandaloneYears
+        if year < thresholdYear {
+          return "\(year) - \(min(thresholdYear, year + 10) - 1)"
+        } else {
+          return sectionIndexTitle
+        }
     }
   }